調整機車費率欄位格式為int

+++ Domain/TB/ProjectAndCarTypeDataForMotor.cs
@@ -85,18 +85,18 @@ namespace Domain.TB
         /// <summary>
         /// 計次分鐘單位(幾分鐘算一次錢)
         /// </summary>
-        public double InsuranceMotoMin { set; get; }
+        public int InsuranceMotoMin { set; get; }
         /// <summary>
         /// 計次金額(每分鐘單位多少錢)
         /// </summary>
-        public double InsuranceMotoRate { set; get; }
+        public int InsuranceMotoRate { set; get; }
         /// <summary>
         /// 安心服務欄位預設值(0:無法購買, 1預設不勾, 2預設勾選)
         /// </summary>
-        public double InsuranceDef { set; get; }
+        public int InsuranceDef { set; get; }
         /// <summary>
         /// 機車安心服務基消分鐘數
         /// </summary>
-        public double BaseInsuranceMinutes { set; get; }
+        public int BaseInsuranceMinutes { set; get; }
     }
 }
